+++ docs/changelog.md
@@ -18,6 +18,7 @@ title: Changelog
 - **BREAKING CHANGE** remove the `@` tag for component rendering. Use the component name alone.
 - **BREAKING CHANGE** disable `{% <app>.<library>.<component_name> /%}` reference. Only allow `{% [<library_name>.]<component_name> / %}`
 - **BREAKING CHANGE** rename `@v` tag to `livevar`
+- **BREAKING CHANGE** `block` inside components renamed to `slot`
 
 ### Added
 - `request.tetra.current_url_path` that holds the path without query params